fix: colocate complete CRD exports in each v1alpha1.ts source file
Move complete V1CustomResourceDefinition exports (packageCRD,
exemptionCRD, clusterConfigCRD) from definitions.ts into their
respective v1alpha1.ts files so crd-manifests can target individual
source files directly. definitions.ts becomes a re-export barrel.
