+++ src/pepr/operator/crd/sources/cluster-config/v1alpha1.ts
@@ -3,7 +3,11 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later OR LicenseRef-Defense-Unicorns-Commercial
  */
 
-import { V1CustomResourceDefinitionVersion, V1JSONSchemaProps } from "@kubernetes/client-node";
+import {
+  V1CustomResourceDefinition,
+  V1CustomResourceDefinitionVersion,
+  V1JSONSchemaProps,
+} from "@kubernetes/client-node";
 
 export const v1alpha1: V1CustomResourceDefinitionVersion = {
   name: "v1alpha1",
@@ -145,5 +149,22 @@ export const v1alpha1: V1CustomResourceDefinitionVersion = {
       },
       required: ["spec"],
     },
+  },
+};
+
+export const clusterConfigCRD: V1CustomResourceDefinition = {
+  apiVersion: "apiextensions.k8s.io/v1",
+  kind: "CustomResourceDefinition",
+  metadata: { name: "clusterconfig.uds.dev" },
+  spec: {
+    group: "uds.dev",
+    scope: "Cluster",
+    names: {
+      plural: "clusterconfig",
+      singular: "clusterconfig",
+      kind: "ClusterConfig",
+      listKind: "ClusterConfigList",
+    },
+    versions: [v1alpha1],
   },
 };
